+++ core/src/commonMain/kotlin/io/islandtime/YearMonth.kt
@@ -1,8 +1,7 @@
 package io.islandtime
 
+import io.islandtime.base.DateTimeField
 import io.islandtime.internal.*
-import io.islandtime.internal.appendZeroPadded
-import io.islandtime.internal.toIntExact
 import io.islandtime.measures.*
 import io.islandtime.parser.*
 import io.islandtime.ranges.DateRange
