+++ core/src/commonMain/kotlin/io/islandtime/parser/DateTimeParserDsl.kt
@@ -1,6 +1,6 @@
 package io.islandtime.parser
 
-import io.islandtime.DateTimeField
+import io.islandtime.base.DateTimeField
 
 @DslMarker
 annotation class DateTimeParserDsl
