+++ core/src/commonMain/kotlin/io/islandtime/ranges/DateTimeInterval.kt
@@ -2,7 +2,9 @@ package io.islandtime.ranges
 
 import io.islandtime.*
 import io.islandtime.MAX_DATE_TIME_STRING_LENGTH
+import io.islandtime.base.DateTimeField
 import io.islandtime.measures.*
+import io.islandtime.measures.internal.minusWithOverflow
 import io.islandtime.parser.*
 import io.islandtime.ranges.internal.*
 import kotlin.random.Random
