+++ core/src/commonMain/kotlin/io/islandtime/ranges/InstantInterval.kt
@@ -3,9 +3,10 @@ package io.islandtime.ranges
 import io.islandtime.*
 import io.islandtime.MAX_INSTANT_STRING_LENGTH
 import io.islandtime.appendInstant
+import io.islandtime.base.DateTimeField
 import io.islandtime.measures.*
-import io.islandtime.operators.endOfDayAt
-import io.islandtime.operators.startOfDayAt
+import io.islandtime.endOfDayAt
+import io.islandtime.startOfDayAt
 import io.islandtime.parser.*
 import io.islandtime.ranges.internal.buildIsoString
 import io.islandtime.toInstant
