+++ core/src/commonMain/kotlin/io/islandtime/ranges/TimePointInterval.kt
@@ -1,7 +1,8 @@
 package io.islandtime.ranges
 
-import io.islandtime.TimePoint
+import io.islandtime.base.TimePoint
 import io.islandtime.measures.*
+import io.islandtime.measures.internal.minusWithOverflow
 import io.islandtime.ranges.internal.*
 import kotlin.jvm.JvmName
