+++ core/src/commonMain/kotlin/io/islandtime/ranges/ZonedDateTimeInterval.kt
@@ -1,9 +1,10 @@
 package io.islandtime.ranges
 
 import io.islandtime.*
+import io.islandtime.base.DateTimeField
 import io.islandtime.measures.*
-import io.islandtime.operators.endOfDayAt
-import io.islandtime.operators.startOfDayAt
+import io.islandtime.endOfDayAt
+import io.islandtime.startOfDayAt
 import io.islandtime.parser.*
 import io.islandtime.parser.expectingGroupCount
 import io.islandtime.parser.throwParserFieldResolutionException
