+++ core/src/commonMain/kotlin/io/islandtime/ranges/internal/Common.kt
@@ -1,7 +1,7 @@
 package io.islandtime.ranges.internal
 
 import io.islandtime.measures.*
-import io.islandtime.measures.minusWithOverflow
+import io.islandtime.measures.internal.minusWithOverflow
 import io.islandtime.ranges.TimeInterval
 
 internal fun secondsBetween(
